(GH-68) Search by namespace in Get-ForgeModuleInfo
This commit adds the ability to return all modules in a given namespace
on the forge with the `Get-ForgeModuleInfo` public function, replacing
the functionality previously found in the `Get-ForgeDscModules`
function in the util script.

Calling `Get-ForgeModuleInfo` without any parameters now returns all of
the Puppet modules in the DSC namespace by default (with a pagination of
5) and only specified modules if the `Name` parameter is used.

Updates tests and documentation for new functionality.