(GH-68) Refactor Publish-PuppetModule
This commit refactors the `Publish-PuppetModule` public function to
include the enhancements from its reimplementation in the Util script
as `Publish-PuppetDscModule` and to call the new `Export-PuppetModule`
public function when needed.

This commit also updates the function to use the `FORGE_TOKEN`
environment variable as a default forge token if one is not specifically
passed.

The updated function is a nearly full-rewrite of the original
implementation with updated tests and documentation.