(GH-68) Add `Get-UnpuppetizedDscModuleVersion`
This commit moves the `Get-UnpuppetizedDscModuleVersion` function from
the Util script into the module as a public function, adding tests and
documentation.

In the process, it refactors the function for some UX improvements,
enabling users to specify the PowerShell repository to check as well as
the Forge URI and NameSpace to compare against. It also adds two options
for filtering the unpuppetized versions - `OnlyNewer`, which returns
only the versions found in the PSRepository for each module which are
newer than the newest version on the Forge; and `MinimumVersion`, which
filters out any versions older than the version specified.

In the process of implementing this function, it was necessary to
refactor the `Get-PowerShellDscModule` function to take the `Repository`
parameter for searching repositories other than the default. Further,
this function was refactored to write verbose messaging for the queries
being executed against the PowerShell repository and to add some error
handling to ensure that errors for failed queries are emitted without
stopping processing *and* only successful queries write output to the
pipeline.

In the process of implementing `Get-UnpuppetizedDscModuleVersion`, it
was also necessary to modify the `Get-ForgeModuleInfo` function to
improve error handling, messaging, and logging for the same reasons as
with `Get-PowerShellDscModule` above.