(MAINT) Make Get-UnpuppetizedDscModuleVersion more performant
This commit modifies the Get-UnpuppetizedDscModuleVersion function more
performant by piping the output of Get-PowerShellDscModule to the process
block for handling the version comparison logic instead of having to wait
for it to finish entirely; this is handy when calling it without any names
because otherwise it has to wait to enumerate the entire Gallery for any
modules with DSC Resources.

As this is happening prior to any release, no update to the changelog is
required for this patch.