(CONT-257) Add PDKTemplateRef parameter
Prior to this commit it was not possible to specify the version of
pdk-templates that would be used when puppetizing a Dsc module.

PDK help sugegsts that the default ref is 2.5.0, however in practice it
would pull down the latest tag from the pdk-templates repository. At the
time of commit, this was 2.6.0.

The above diff causes some issues when puppetizing a resource due to gem
incompatibilities.

This commit adds a new parameter called PDKTemplateRef to
`New-PuppetDscModule` and `Initialize-PuppetModule` so that users can
optionally provide a different pdk-templates version.

The default value has been set to 2.5.0.