(GH-240) Fix incorrect enum generation
Prior to this change enums with a single value would be generated with a
lowercase counterpart but the enum definition was malformed.

For example, given the the value `Yes` the following enum would be
generated:

`Optional[Enum['Yes' 'yes]]`

This was happening because when a PowerShell array has a single value it
will output a string.

This meant that when being assigned the lowercase variant of the value a
string concatination occured causing the join on the next line to be
ineffective.

This change updates the flow to use a `System.Collections.ArrayList` so
that we will always have the expected datastructure.

After this change, given the value `Yes' the following enum would be
generated:

`Optional[Enum['Yes', 'yes]]`

This is expected syntax and does not cause Puppets syntax parser to
raise an error.