(CAT-1322) - Only collect forge modules if no module passed to puppetize
Prior to this commit, the puppetize workflow would always make a request
to the forge API for the top downloaded DSC modules, despite if they
were to be puppetized or not.

Now the puppetize workflow will only make a GET request to the forge API
if no module was supplied as an input to the puppetize workflow.

This means that there is reduced execution time for workflow runs that
have supplied a module to be puppetized.

Signed-off-by: jordanbreen28 <jordan.breen@perforce.com>