(CAT-1484) - Add ensurable property to resource
This commit adds a ensurable property to each resource which does
not already have an ensure property defined. This has a couple of
beneifts:

1. Prevents issues as seen in CAT-1484, where puppet namevars and DSC
   keys differences arise. By default, puppet will ignore namevar
   changes if they are the only atrtibutes passed in a manifest. This is
   a design of puppet and makes sure as namevars should only identify a
   resource. DSC's key values behave differently to this.
2. Will now clearly show in docs that the resource CANNOT be ensured.

The property sets a default of 'false' and will only accept values of
'False' or 'false'. The value will not need to be specified in
manifests.