Add DRY design decision
This commit contains the reasoning behind the bulk of the changes in this PR.  Although the primary aim of this PR was to fix the broken acceptance tests.  I took the opportunity to dry up a bunch of duplicated github action code.  This ADR goes into more detail about the 'why'.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>