(CAT-1869) - Add dsc_timeout param to all dsc resources
This commit adds the dsc_timeout param to all dsc resources generated by
Puppet.Dsc. This allows the user to set a configurable amount of time,
in seconds, in which puppet will wait for the dsc invocation to finish.

If the timeout passes, and DSC has not returned, the powershell process
defined in puppetlabs-pwshlib's dsc_base_provider, will quit the
operation and move on.